Related to issue #632; implements Survivor.Pilots as such: - Defaults to 0 if Crewed=no - Defaults to 1 if Crewed=yes
NOTE: I am not sure if there is a check for Crewed=yes before SpawnSurvivors() is called; 
if yes, the additional check for Crewed before executing the pilot survivor routine I added is superfluous;
if no, this additionally fixes a bug where having a positive PilotChance would have spawned pilots even on Crewed=no vehicles.

The placement code of the actual spawned pilots is unmodified; it is nearly identical to the passenger one, so I assume it'll work for multiple pilots. If not, that'll need fixing.

git-svn-id: svn://svn.renegadeprojects.com/ares/trunk@221 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd