Something like two thirds of the implementation of issue #342. I'm pretty sure the specific operator check will not work/compile the way it is, and it's lacking the check for being above a building - the rest should be fine. Will finish once I get D's advice on this.
It is/will be implemented as Operator= on any TechnoType, where the value can be either _ANY_ to indicate that it doesn't matter who's operating, as long as somebody is, or any one InfantryType (emphasis on one - this flag does not take a list).

Examples:
Operator=_ANY_
Operator=CLEG

Reading the string "_ANY_" takes precedence over InfantryType reading, so an InfantryType [_ANY_] will not work with this flag.
But if you try something like that, you're a moron anyway :P

git-svn-id: svn://svn.renegadeprojects.com/ares/trunk@234 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd