Another mid-work commit of several weeks of work: - PassThrough logic should work now (issue #663, issue #667) - Raidable buildings might work (issue #665) - Most of Advanced Rubble should be there, but reconstructing the rubble is not entirely hooked up yet (issue #664) - Most of trench traversal is there, but, as with Advanced Rubble, it's not hooked up yet (issue #666)
The parts not hooked up yet are not yet integrated because, currently, no facility exists to control the mouse cursor and add new actions, which means, in both cases, the actual functions performing the necessary actions exist, but the part where it actually triggers them to be executed when the user clicks is missing.

Advanced Rubble is roughly 95% there, the only cases not working yet (as far as I can see) are:
- Reconstructing when the rubble is undamaged; since you have to reconstruct the rubble with an Engineer, if you can't enter the rubble, you can't reconstruct.
- Not losing the Engineer; it is planned to not lose the Engineer through reconstruction, but that part is not coded yet. Currently, the Engineer will vanish when entering the building, as usual.

Raiding buildings in itself should work, but the custom code has no special handling for ownership; if the game's own code correctly turns over ownership as it does with urban combat buildings, this system <i>should</i> be fully operational. If it does not, code will have to be added to handle the temporary ownership transfer (if only so the raiding party can actually leave the building again).

And, of course, as usual, D will probably find a bunch of fixable things.
So this is very much a work in progress, but a lot of progress has been made.

git-svn-id: svn://svn.renegadeprojects.com/ares/trunk@245 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd