<b>Issue #664: Advanced Rubble</b> The BuildingType specified under Rubble.Destroyed is now hardcoded to have the following properties: Capturable = false; TogglePower = false; Unsellable = true; ClickRepairable = false; CanBeOccupied = false;
In addition, the rubble will be spawned with 99% of its nominal strength; combined with not being ClickRepairable, this ensures that Engineers will always get a repair cursor over rubble. (The alternative to this hack would be having to hack the UI, mouse and event logic just for one spot for this one logic.)
Besides, having the scorched rocks of an ex-building be ClickRepairable doesn't make much sense in the first place. The other settings follow the same sentiment.

Lastly, Engineers should now not be lost anymore when they repair rubble. Please keep in mind that that means THE ENGINEER KEEPS EXISTING.
Which means it takes up space.
Which means it blocks a cell.
Which means if you moronically decided to have the intact building have a larger foundation than the rubble, the game will try to place the building on top of the Engineer.

If you get IEs out of that, that's your fault.

<b>Issue #665: Raidable Buildings</b>
Raidable buildings should now correctly change ownership when raided, and return to their original owner once emptied.
Raided buildings are unsellable, that is, the temporary owner of the building cannot sell it.

git-svn-id: svn://svn.renegadeprojects.com/ares/trunk@258 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd