<strong>Issue #666</strong> Added line-drawing/linking between allied structures for Ares's linking system, as discussed in thread #1424.
<strong>Issue #699</strong>
This was an oversight when fixing up the CanBeOccupied check for issue bug:665 and issue bug:218.
Since the original game limited occupation to the same owner, it didn't have to check whether there were already people in there or not, only if the building was
- yours or neutral
- not in red hp
- not full
- not mind controlled
We removed the check for ownership, and instead added a check for raidability.
So what does the game see when Player B wants to occupy a Player A-occupied, raidable building?
- It's not mine, but it's raidable
- It's not in red
- It's not full
- It's not mind controlled
...so it's free to be occupied! Yay! -_-

Since the code kinda-sorta suggests the same problem exists with urban combat buildings, I have proactively applied the fix to both raiding and normal UC occupation.
This should fix issue #699 - if the fix interferes with urban combat, it can easily be limited to only raiding.

<strong>Issue #700</strong>
The PassThrough logic didn't trigger re-evaluation of the raiding status.
The necessary function already existed in the raiding logic, it just wasn't called by PassThrough.
This should fix issue #700.

git-svn-id: svn://svn.renegadeprojects.com/ares/trunk@266 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd