This should fix the bug mentioned by Marshall in the comments of issue #698. The part deciding "is this a raid?" only checked whether the owners were different and if the building was already raided, not checking whether there was actually any <em>need</em> to raid the building. After this patch, the raiding logic should only fire on buildings not belonging to houses with MultiplayPassive=true.
(Either that, or I just broke the complete occupying logic >_>)

The main "bug", if I interpret everything correctly, is simply that the function to change the owning house automatically triggers the "Building captured!" sound.
Should be fixable through an exception somewhere.
As indicated in the bug, that fix is scheduled for 0.2, since it doesn't affect the functioning of the logic itself, and is, to a certain extent, correct - one <em>does</em> capture the building when one raids it. Just not in the way the game thinks one did.

git-svn-id: svn://svn.renegadeprojects.com/ares/trunk@316 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd