<strong>Fixed issue #1350</strong> - AttachSound should be fixed now.
As the documentation says, DestroysBridges doesn't change the logic of whether bombs can be attached to the huts, only whether they will destroy the bridge when they explode. Only infantry with Ivan=yes can attach bombs to huts normally, and fixing that will take time. Make a new issue for that one...
Related to issue 1350 .

git-svn-id: svn://svn.renegadeprojects.com/ares/trunk@897 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd