<strong>Fixed issue #1037</strong> - <pre><code>;artmd.ini [Animation1] CustomPalette=abcd.pal ; always use abcd.pal
[Animation2]
CustomPalette=lib~~~.pal ; use libtem.pal, libsno.pal, etc. "~~~" is replaced with the theater's three-letter tag you should already be familiar with</code></pre>

Note that as a convenience, the ~~~ is now supported in other contexts where Ares' Palette reader is used, that is, on TechnoType/SW cameos. It's <strong>not</strong> supported for TechnoTypes themselves, as that would require a lot more clusterfucking to handle remapping colours and whatnot.

TODO: store palettes in a hash of some kind to avoid loading the same palette file N times.
Related to issue 1037 .

git-svn-id: svn://svn.renegadeprojects.com/ares/branches/issue-340-prism-forwarding@938 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd