Bounty logic implementation and first binary. Graion's and Joshy's code. Related to issue #203, Related to issue #1523.
Documentation:
[TechnoType]->Bounty.Modifier (float, defaults to 0)
Multiplier of the enemy victim's Cost/Damage which should be granted as bounty.
[TechnoType]->Bounty.Message (bool, default's to no)
If set to yes, the granted money will be written out above the unit.
[TechnoType]->Bounty.FriendlyModifier (float, defaults to 0)
Multiplier of the friendly victim's Cost/Damage which should be taken as bounty.
[TechnoType]->Bounty.FriendlyMessage (bool, defaults to no)
If set to yes, the taken money will be written out above the unit.
[TechnoType]->Bounty.Pillager (bool, defaults to no)
If set to yes, the unit will grant money for the Damage it delivered and not when it destroys a unit.
[TechnoType]->ImmuneToBounty (bool, defaults to no)
If set to yes, the TechnoType can't grant money.

git-svn-id: svn://svn.renegadeprojects.com/ares/branches/ft-bounty@1045 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd