<strong>Related to issue #1362</strong>:
ChangeOwner polished to meet with quality standards.

Two new tags are also in:
[TechnoType]->ImmuneToAbduction=boolean, defaults to false.
Specifies if the unit can be abducted.

[Weapon]->Abductor.AbductBelowPercent=float, defaults to 1.
Specifies the percentage of health, in below the target can become abducted. I mean, if it says 0.5, then the targget can only be abducted if it's already lost at last half of it's health.

git-svn-id: svn://svn.renegadeprojects.com/ares/branches/issue-1362-abductor-fixes@1124 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd