Added two new functions to Prism Forwarding which maintain the connection between SlaveTower->SupportTower and MasterTower->Senders. Whenever the SupportTower is changed, the supporting tower is removed from the original master tower's Senders list and then added to the new master tower's Senders list.
RemoveAllSenders() removes all registered senders using SetSupportTarget().

Still to do: SupportingPrisms is said to be the longest backward chain, but it is still decremented when removing a tower without any range checks. Shouldn't it be set to the longest backward chain of all slave towers still registered in the Senders list plus 1? And shouldn't this value be updated when adding a tower also?

(Searched for easter eggs in the code, but didn't find any.)