Related to bug #896113: Infantry being created @ 0,0,0, cause uncertain
This bug should be fixed. CellClass::FindInfantrySubposition() was the culprit. It converted valid coords into 0,0,0 if the cell was occupied and no more infantry would fit. EjectSurvivor() then made a mess out of it, because it didn't verify whether the cell the coords pointed to are actually valid. The game always returns a cell object and even updates it's fake location on the map so everything appears to be fine, even if it is not.