Related to issue #1000839: Random crash on mission reload
Ares cached the ColorScheme to be used for single player loading screens. The color schemes were reset in the middle of loading, and the old pointer would point to invalid data, which could lead to a crash.

Now, the array index of the selected ColorScheme is cached and resolved to the correct ColorScheme when it is needed.