Updated the version information and resources.
The numerical file version stays in the recently introduced format <year>.<dayofyear>.<minuteofday>.0. The numerical product version will be the same for beta binaries. For release binaries, the product version will take the form <major>.<minor>.<patch>.0 (like 0.2.1.0 for a hypothetical 0.2p1).

The string file version ("FileVersion") is a string representation of the numerical file version. For beta binaries, "ProductVersion" will be the same. For release binaries, "ProductVersion" will be the human-readable display name of the version name (like "0.2p1").

To distinguish between beta and release binaries, check for the VS_FF_PRERELEASE flag. It will set for betas and clear for release binaries.