Added the new tag [GlobalControls]DebugKeysEnabled= (boolean, defaults to yes). If it is set to no, the AI Base Plan, AI Control, Dump Types and Map Snapshot keyboard commands are disabled. If they are executed, the language string "TXT_COMMAND_DISABLED" is shown instead.
This tag can be used to prevent users from gaining an unfair advantage in multi-player games, from revealing vital secrets, and from deliberately causing a Reconnection Error in release versions of a mod.

TXT_COMMAND_DISABLED can include one "%s", which is replaced with the actual command name. In Ares, this string defaults to "The %s command is not available.".