Updated forEachObjectInRange() method.
It now returns negative values if the input is invalid. Otherwise it always returns the count of objects that successfully invoked the callback action.

These changes do not affect how the function behaves internally, and no caller uses the return value for anything yet, so this is actually just a minor service update.