Added forEachObjectInCellRange and changed the Psychic Dominator to use it.
forEachObjectInCellRange checks the cell distance to the center cell instead of each object's distance to the center. This means, if a cell is close enough, all objects on it will be affected, regardless of their sub-position on the cell.

This fixes a problem with the original Yuri's Revenge matches (including single player missions) featuring the Dominator, where it would not capture all units in a 3x3 area (only if defaulting to DominatorCaptureRange=). Instead, only the target cell and its four direct neighbors would be affected. The diagonal cells were too far away for a default radius of 1.

For instance, freeing Einstein gives you a one-shot PsyDom, which would not be able to capture all nine units waiting in Yuri's base. You would get only five units - three of them Gatling Tanks - which makes this mission harder. All missions where the PsyDom is used against you became easier, naturally.