Split forEachObjectInCellRange into two methods. Replaced the almost-cellspread-like distance calculation by a better version based on code by DCoder. It matches the first 10 cell spread areas, and only misses the last one because WW fucked up.
forEachCellInCellSpread invokes a callback on each cell in an area determined by the CellSpread logic. This does not use the CellSpread offset table, so it will work for values greater than 10.

forEachObjectInCellSpread uses the above method to get all cells in the area and then invokes a callback function on each object on that cell.