Related to issue 1024707: Multi-Engineer related Reconnection Error.
MultiEngineer is always considered enabled, except for skirmish games. This prevents the setting from causing REs if one player played a skirmish game with ME enabled and one player didn't enable ME. In skirmish it's still possible to disable MultiEngineer if the checkbox in the skirmish menu is allowed to be shown.

Minor change: comment edited.