Related to issue 1024127: Multi-Engineer won't work in campaigns.
Skipped the check for the game mode. This makes ME work in single-player missions. Also changed the way tech structures are detected. They have to belong to Neutral or Special initially. It is not sufficient any more to give them to a MultiplayerPassive house, because missions might define their own houses.