Related to issue 1025156: Player drop down colors affects in-game messages
As long as there is no dedicated tag to change it (which won't happen in 0.2) the message color for map action 11 messages is hard coded to the original [Color] items. Customizable drop down colors affected the colors, which is an unintentional change, in other words, a bug.

Message colors are hard coded to color 10 for side 1, color 5 for side 2 and color 12 for every other side, just like in the original YR without the customizable drop down colors.