Related to issue 1025226: PoweredBy weirdity
EMP_Threshold was used in a weird way. Comment said it's checking for the "inair" value, but it checked for all negative values. "inair" is equivalent to -1. Negative EMP threshold destroys a unit only if it is in air. Positive threshold will destroy the unit regardless of its height above ground. Thus, positive threshold will also destroy units in air, so it should be checked also.

Note: EMP was changed not to use InAir() because units taking off and landing would be misinterpreted. Instead, it currently checks for GetHeight() > 0. I won't reproduce this there, because I can't rule out that this would break something.