Related to issue 1002020: Stale Mind-Control pointers
The problem is this: If a mind-controlled unit is destroyed, the mind-controller gets notified and removes the victim from its internal list the hard way - the pointer is considered invalid and won't be accessed. Thus the back-linking field isn't reset.

The dead victim keeps its valid link to the controller and plays its death anim, keeps sinking or crashing.

If the mind-controller is killed now and finishes its death anim earlier than the victim, it is uninitialized and freed earlier than the victim also. At this point, the link from the victim to the controller becomes invalid. The crash happens the moment the game checks whether the mind control link should be drawn.

If the controller dies before the victim, the mind control link is removed by the real unlink function which also checks and updates the victim properly.

The applied solution removes the pointer from the victim to the mind-controller before AnnounceInvalidPointer is called.