Related to issues 1002020, 896263, and 895954: Dead Mind-controllers stink
Implemented the mind-controller link differently. Instead of clearing the pointer on the object when the object is about to be destroyed, it is now cleared like any other pointer. That is, the object is notified that the mind controller became invalid and the pointer is reset immediately.

This should fix the "mind-controller finishes death animation before victim sunk", as well as problems with map actions which temporarily remove units from the map when placing buildings.