Related to issue 896366: Advanced Rubble causes crash at start of game
Added the missing check for a valid "Rubble.Destroyed" building type ID. Previously, if the tag was set but it didn't resolve to any known building type, Ares would still try to set properties on this type.

Now it is possible to disabled Rubble.Destroyed in game modes or maps. Still missing: Logging for INI parser errors.