Issue #1050572: Lag because of Firewall handling.
Whenever a techno is asked about its firing state, it tries to find out whether the direct line to its target contains a Firestorm wall section. This was done even if no house had an active Firestorm. While firing it checked the same path again, but would abort if there is no Firestorm active.

The contained fix checks whether there is at least one active Firestorm for both functions.

Additionally, it does not iterate over all houses for every techno every frame. Instead, the value is cached and only updated if a house changes the Firestorm state. This should reduce the number of calculating the projectile path from O(N technos * M houses) to an amortized O(1), not including the time saved by not checking the projectile path for Firestorm wall sections if none are active.