Related to issue #895788: Parachute vehicles yet again
Remember whether a unit was put in the high part of a cell instead of the usual ground. When the unit is removed, it is removed from where it was put, not from somewhere else.

This fixes the issue of cells becoming impassable once a unit drove over a bridge above it.