Implemented a cache for dynamic CSF labels.
NOSTR and MISSING texts are added to the label list whenever they are encountered. If the same string is encountered many times, it is added as many times to the list. This happened for every match.

Now the texts are added once to a hash map instead. This replaces excessive (in terms of "uses more than needed") memory usage by the cost of logarithmic (instead of constant) runtime.