Implemented EVA.Tag=
All new EVA types have to be named in the new [EVATypes] list in evamd.ini. This is a list like every other list, i.e. the part left of the equal sign is ignored. The voices called "Allied", "Russian" and "Yuri" are always added.

Example:
[EVATypes]
0=Cabal
1=GDI

Now, these values can be used as keys in the EVA sound sections to point to a EVA-specific filename.

Example:
[EVA_LowPower]
Cabal=ccab053
GDI=cgdi053

To give a side a certain EVA voice, use [Side]EVA.Tag=. It takes an EVA type from above, or one of the values "Allied", "Russian", "Yuri" or "none".