Restore Map actions for the Firestorm defense SW
Map action 92 "Activate Firestorm": Enables the Firestorm SW for the house owning the trigger, regardless of charge state. The firestorm will not drain the charge, but it will shut down when the house is low on power.

Map action 93 "Deactivate Firestorm": Shuts down the Firestorm SW for the house owning the trigger.

Both actions use none of the parameters. Parameters should be set to "0,0,0,0,0,0,A".

It is recommended to not use these actions for human player houses as long as they have access to the Firestorm SW.

Untested.