Related to issue #894978: Restore the "CloakStop" function
CloakStop has been fixed. It failed, because an additional check for a techno's Cloakable member overwrote the result of the IsCloakable() method.

Additionally, Is_Moving_Now() has been replaced by Is_Moving(), because it doesn't check the current speed in case the unit is turning, thus the techno stays uncloaked.