Made the dimming effect for deactivated voxel units customizable
[AudioVisual]DeactivateDimEMP= (float, defaults to 0.8)
[AudioVisual]DeactivateDimOperator= (float, defaults to 0.65)
[AudioVisual]DeactivateDimPowered= (float, defaults to 0.5)

The first applicable value is used, the other values will not be considered. That is, if an unpowered unit is under EMP, the DeactivateDimEMP is used. This is purely cosmetic and the order may change in the future. A value of 1.0 disables the dimming effect. Supported range is 0.0 to 1.0, though higher values might also work to brighten the unit.

This only affects UnitTypes and only voxels.