Reimplemented Tiberium Heal feature
The feature works the same way as in Tiberian Sun.

[General]TiberiumHeal= (double, defaults to 68/60 = ~1.13333, once every 68 seconds) The interval between healing steps being applied.

[TechnoType]TiberiumHeal= (boolean) Can this techno heal if standing on a cell containing Tiberium? Defaults to no.

The health restored depends on the type; see IRepairStep and RepairStep.

Ares also supports the TIBERIUM_HEAL veteran ability, which can override [TechnoType]TiberiumHeal=.

When technos that have TiberiumHeal=yes are destroyed, they leave behind a random amount of tiberium. The veteran abilities are not considered here.