Implemented tiberium damage for infantry types
[InfantryType]TiberiumProof= (boolean, defaults to no) Whether the infantry is immune to tiberium damage.

This feature also supports the veteran ability TIBERIUM_PROOF.

The damage delivered when entering this cell is a 10th of the tiberium's Power= value.

The commit contains code to create a small Visceroid, but it is commented out because the wandering logic is not complete yet. Differences to the original logic are the validation of the SmallVisceroid tag and the deletion of the unit if it could not be placed (which should never happen).