SensorArray fixes
Buildings with SensorArray=yes use SensorsSight= to detect stealth and subterranean units, as in Tiberian Sun.

SensorArrays were not supposed to be powered, to change owner (by capture or mind-control), and weren't updated for temporal weapons. All of this should work now as one would expect.

The bug that the sensor effect would stay active even if the array was destroyed is fixed. It was because SensorsSight was used to activate the effect, while CloakRadiusInCells was used to remove it. If the latter was smaller, the bug happened.

The logic was also changed to not add the same array more than once (same for removal).

Newly added EVA events, which should be self-explanatory: EVA_CloakedUnitDetected and EVA_SubterraneanUnitDetected.