Storage extended
[BuildingType]Refinery.CanSpill= (boolean, defaults to no): Whether the building will spread its stored tiberium when destroyed.

[BuildingType]Refinery.UseStorage= (boolean, defaults to no): Whether the building will store the tiberium instead of converting it to money directly. Tiberium that is converted directly does not require storage space.

Bonus money created by ore purifiers are always stored on the bank account instead of in the silos. This is because the bonus is calculated in tiberium bails, which would count against storage. When the silo is destroyed, this bonus money would be placed on the map, which would create tiberium out of nothing.