Related to issue #1051260: AI team building issue
Converted the ugly macros into a half-as-ugly template function. The important change is swapping the operands in a comparison (line 52 old, 39 new). It tried to find the higher value, but TeamCreationFrame was already initialized with the highest value a signed int can hold, thus, it didn't do anything.

Thanks to DCoder for finding this out.