Related to issue #1075936: Powered aircraft types still exploding permanently
The PoweredUnitClass updated its state even if the unit was crashing already. This caused aircraft to explode every update step anew. Update is disabled for destroyed units now.

Because the method to disable the unit contained a code path that destroys flying aircraft, it was possible that the unit's memory (ExtData) was freed while currently being used. This code has been removed. Instead, the return value is used to indicate the unit has to be destroyed.