Also support Insignificant and DontScore for map actions 60 and 61.
The counter used in the previous approach does not keep track of technos marked as Insignificant=yes and/or DontScore=yes. Now, the old way is reimplemented that iterates all technos and checks their types. Ares will always chose the fastest possible approach. The techno type lookup is optimized in both versions.