Parser class added
Parser<T, Count> class allows to parse a C string into Count elements of T.

Parse() reads elements until Count is reached or a non-parsable value is encountered. It returns the number of successfully read elements. Following elements in the output buffer are unchanged.

TryParse() reads Count elements into a temporary buffer and only copies them to the output, if Count values could be read. It returns a bool to indicate this situation.