Reimplemented INI_EX using the new Parser template
Change of behavior: Previously each reader function just reported back whether the tag was declared. Thus it reported success even if the format was not valid and the value not could be parsed. Now, success is only reported if the tag is declared and its value can be parsed. (This implicitly fixes error logging for Valueable<T>. It only happens when the reader reports failure and the tag was declared, which could not happen at the same time. If and only if the reader returned failure, the tag wasn't declared.)

Change of behavior: Read2Integers(), Read3Integers(), Read4Integers() and Read3Bytes() report failure if not all values could be parsed. Previously, they reported success even if not a single value could be parsed. The buffer is still partially filled. (Because Valueable<T> uses buffers, this will now cause partial values not to be accepted any more.)

Change of behavior: Each element of Read2Integers(), Read3Integers(), Read4Integers() and Read3Bytes() is parsed separately now, each one supporting the formats reading a single element would support. Single integer format can (and could) be decimal "19", hexadecimal "13h" or decimal with a prefixed dollar sign "$19". But previously, two, three or four integers only supported decimals.

Change of behavior: When multiple elements are to be read, the value is split at the comma signs (as it was previously). Now each of these parts is then trimmed of all whitespace again. Previously, this was not supported. For example, "19, 12 , 5" is now valid where it previously wasn't.