Update for #include handling
If Includes::LoadedINIs.Count in the condition is 0, there is no need to delete LoadedINIs. There aren't any, because the count is already 0.

Changed "j > 0" to "j >= 0" to also remove the last item's name, otherwise it would be left in the list and the file would not load any more, because it would always count as "already loaded". The file name would always be "ra2md.ini".