Related to #1029055: Crash on loading screen
The special name text pointer cannot be null, because the value is passed as source to wcscpy, which is not allowed to be null. In the original game this happens when the country index is still set to random. In Ares it also happened if the ExtData wasn't available and the country index was higher than 9 (because of the default null value) or 0 (because of the broken test instruction fixed recently).

Now, the special name pointer points to an empty string in cases where the game would use a null pointer.