File.ObserverFlag and File.ObserverBackground update
The read value is now automatically converted to lower case. The PCX loader seems to require this.

If a SHP or PCX file cannot be opened, the value is cleared and not parsed again next time an INI is read.

If the PCX image is to be used, the loaded SHP is cleared.

The File.ObserverBackground PCX was never loaded. This was ok if the file was loaded by some other part of the code. Otherwise the file could not be drawn.