Related to #1171031: Custom foundations on minimap
Custom foundations did not show up on the minimap, because in an unpatched location the foundation type index was used as an index into an array which defines the pixel offsets to be drawn. Because Ares uses a dummy value of 0x7F to mark custom foundations (which was way beyond the pixel array size of 22), some arbitrary data related to cameo lists was read.

For now, custom foundations are hard-coded to draw like 2x2 foundations do on the minimap.