The foundation invalidator had some issues.
Same reasons as the previous commit. Register check didn't what one might think it does. Wrongfully accepted blank values. Also logs actual tag name now.

Note: The only acceptable new value is "Custom" because Ares handles it. Yet, this function needs some more work because Ares sets the foundation tag to a sentinel value, which is used as index for the default value. It currently points in the middle of a string.