Pip Distortion Bug fixed
Pip= and OccupyPip= will no longer shift their values by one or two if not defined in an ini. The previous value was used as an index into a list to fetch the current string representation; this was used as default value when reading the tag. Because the list doesn't start with a value of 0 and there's no item with a value of 6, the index never actually equaled the value. This lead to the wrong default string being used. If the tag isn't defined, the wrong string is converted back to a value (which isn't the same as the original value before the read). Aside from that, the largest value was larger than the length of the list by 2, thus the function also read beyond list bounds.

Now the lookup round trip is skipped when the tag isn't defined. The lookup works correctly and never overflows the list.

As a bonus, the tags are parsed as integers when it cannot be resolved to a value. And if that doesn't work, it outputs a debug message.