Template Update 21: Refactored Customizable<T>::Get.
No change in behavior. Aim is to not hand out a local variable, which is required for the next updates.

Value is initialized to T() in the constructor, so T() can be replaced by Value, as long as Value isn't changed without changing Customized or Default.

Set, SetEx and Lock change Value, but they both also mark the Customizable as Customized, thus it would return Value, not T().

BindEx changes Value, but also sets Default to point to Value. The Customizable still counts as not Customized. Previously, Default was set and thus the dereferenced Value was returned. This same logic can now be found in the condition and the true part of the ternary operation.