Related to #1213738: Alpha Lights Crash
AlphaShapeClass does not clear the AttachedTo pointer if it becomes invalid. Instead, it sets IsObjectGone. Ares did not account for this smart move.

The hook AlphaLightClass_UpdateScreen_ShouldDraw broke a test instruction and did not reproduce it. This was most likely the root cause of the crashes (calling DrawVisualFX on a nullptr ExtData).

The hook TacticalClass_UpdateAlphasInRectangle_ShouldDraw exhibited the same behavior and was the actual location of the crash. It was not tested whether the fix mentioned above alone made this crash disappear. Just to be sure, this one is updated as well.

Hooks into the same functions still access AttachedTo without checking whether it is valid, but both updated hooks skip them.