Related to #1204514: Animation warheads and AffectsAllies
Unified all branches of the conditional, the Attacker, SourceHouse and the default one. Default is to affect everything, then Attacker owner or SourceHouse are used to apply either AffectsAllies or AffectsEnemies. The original code is now skipped, and Ares handles it all by itself, hopefully in a way that is easier to read and to overlook.

Updated the comment about neutral targets (or their non-existence) and removed parts that don't apply any more.