Added moddable text to show release note.
Define a string with the label TXT_RELEASE_NOTE in stringtableXX.csf to have Ares display this text in the lower left corner of the game area (next to the Ares stability note and the FPS counter).

Ares.csf should contain an empty TXT_RELEASE_NOTE to be overridden in the string tables, so the text will not appear as MISSING:TXT_RELEASE_NOTE.

Needs testing: If this label is defined in ra2md.csf, it might be overridden by Ares.csf, and thus not display.