Related to #1203221: Cloaking ejects parasites
Parasites aren't thrown out any more when a unit cloaks. They will cloak, keeping the parasite inside.

YR handles parasites manually. The ParasiteClass is informed about invalidated pointers through the victim, not through the owner. The ParasiteClass doesn't get the state of the pointer passed, though; it always sees pointers as being removed from the game. If the victim is handling itself being announced as invalid, it just clears the pointer to the parasite owner, breaking the link.

Now, the ParasiteClass gets the real value and can decide whether it should stay or exit. It only exits when the victim is removed. The parasite owner is only cleared if the victim leaves the battlefield (which it isn't when it cloaks).