Threshold for Explodes on Overlay
[General]OverlayExplodeThreshold= (int, defaults to 0)
Only if the damage to a cell is higher than this does an overlay with Explodes=yes explode. Below this value, the overlay stays intact.

Damage is the unchanged damage value without any Verses applied.

There is no plan to make this per-OverlayType at the moment, but the possibility is there to default to this tag later.