getCellSpreadItems now returns a std::vector.
Instead of returning a pointer to a new'ed DynamicVectorClass, the return type is now a standard vector. Saves us from having to remember to destroy it later. And it is most likely much faster than the old way anyhow, which had to allocate the DVC, and then allocate some more memory for the actual data.