IvanBomb changes/fix
Applying the Ivan Bombs relied on weapon data of other stack frames, which is unsafe and prevents the function being called in different contexts. PlantBomb takes a pointer to a weapon, calls the original function and then adds the customizations.

The BombListClass_Plant have been grouped together. The customization is removed here (see above). The attach sound is silenced globally (customization inserts it back in).

IvanBombSpreadApplicator has been changed to take the weapon type pointer. Note: This only works now because the CellSpreadIterator's slicing problem is fixed.