Related to #1248566: TiberiumSpill and Storage
Reimplemented the Tiberium Spill feature for BuildingTypes. Previously, only spreading the Tiberium was skipped, but a float wasn't popped from the stack. This caused the bug.

The reimplemented version does not take bails one by one and places them on the map. Instead, all bails of a type are removed at once, then spread one by one only if Spill is enabled.