Alloc/Dealloc difference for alpha lights (no bug)
Alpha lights are allocated on the game's heap by GAME_ALLOC, but Ares uses 'delete' to destroy them. 'delete' calls the virtual destructor (with the argument 1); this means the game's 'operator delete' is called after the destructor ran. All well in this case (allocating one the dll's heap would be a bug).

GAME_DEALLOC also calls the virtual destructor (but with argument 0). This skips the call to the game's 'operator delete'. GAME_DEALLOC then goes on and calls the game's 'operator delete' explicitly. No difference, but the name of GAME_DEALLOC makes it clear the alpha light has been allocated on the game's heap.

(Also added some comments that the iterators have been invalidated by destroying an alpha light.)