Incorporate latest knowledge from Stupid Ideas 101.
Returning a const reference parameter is a bad idea because there's no information whether it's a temporary or not (and const references swallow everything). If it is, its life might end after this function returns, yet the return value makes the caller believe everything is ok. This is as good as returning a reference to a local object directly.